requirements.txt
Django==5.2.8
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
redis==7.4.0
dj-database-url==2.3.0
python-decouple==3.8
python-dotenv==1.2.1
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
